% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{assess_recurrence}
\alias{assess_recurrence}
\title{Assess recurrence of covariates}
\usage{
assess_recurrence(dt, id, code, type, rank = Inf)
}
\arguments{
\item{dt}{Data table with patient and covariate information}

\item{id}{Column name for patient identifier}

\item{code}{Column name for codes}

\item{type}{Prefix for output (dx, px, rx)}

\item{rank}{Maximum number of covariates to include}
}
\value{
Data table with recurrence assessment
}
\description{
Assess recurrence of covariates
}
