% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{estBiasTable}
\alias{estBiasTable}
\title{Create 2x2 table for bias estimation}
\usage{
estBiasTable(dt, expo, cova)
}
\arguments{
\item{dt}{Data table with exposure and covariate}

\item{expo}{Column name for exposure}

\item{cova}{Column name for covariate}
}
\value{
2x2 contingency table
}
\description{
Create 2x2 table for bias estimation
}
