% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{hdps-package}
\alias{hdps}
\alias{hdps-package}
\title{High-dimensional propensity score algorithm}
\description{
The high-dimensional propensity score algorithm is a method for
high-dimensional proxy adjustment in claims data. This package implements
the variable transformation and varaible selection parts of the whole
algorithm.
}
\details{
\tabular{ll}{
Package: \tab hdps\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2014-06-05\cr
License: \tab MIT \cr
} ~~ An overview of
how to use the package, including the most important functions ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
\author{
Sam Lendle

Maintainer: Sam Lendle <sam.lendle@gmail.com>
}
\references{
Schneeweiss, S., Rassen, J. A., Glynn, R. J., Avorn, J., Mogun,
H., & Brookhart, M. A. (2009). High-dimensional propensity score adjustment
in studies of treatment effects using health care claims data. Epidemiology
(Cambridge, Mass.), 20(4), 512.
}
\keyword{package}

