% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdps_input}
\alias{hdps_input}
\title{Flexible data input handler}
\usage{
hdps_input(
  data,
  format = c("long", "wide", "matrix"),
  id_col = "pid",
  code_col = "code",
  type_col = "type",
  value_col = "value"
)
}
\arguments{
\item{data}{Input data in various formats}

\item{format}{Input format: "long", "wide", "matrix"}

\item{id_col}{Column name for patient ID}

\item{code_col}{Column name for codes}

\item{type_col}{Column name for domain type}

\item{value_col}{Column name for values (for wide format)}
}
\value{
Standardized data.table
}
\description{
Flexible data input handler
}
