% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdps_interactive.R
\name{hdps_interactive}
\alias{hdps_interactive}
\title{Interactive HDPS Covariate Selection App}
\usage{
hdps_interactive()
}
\value{
Shiny app
}
\description{
Shiny app for interactive covariate selection using HDPS algorithm
}
