% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdps_multi_domain}
\alias{hdps_multi_domain}
\title{HDPS workflow for multiple domains}
\usage{
hdps_multi_domain(
  data_list,
  id_col,
  code_col,
  exposure_col,
  outcome_col,
  n_candidates = 200,
  min_patients = 10
)
}
\arguments{
\item{data_list}{List of data.tables for different domains (dx, px, rx)}

\item{id_col}{Column name for patient ID}

\item{code_col}{Column name for codes}

\item{exposure_col}{Column name for exposure}

\item{outcome_col}{Column name for outcome}

\item{n_candidates}{Maximum number of candidates per domain}

\item{min_patients}{Minimum patients per covariate}
}
\value{
List containing results for each domain
}
\description{
HDPS workflow for multiple domains
}
