% Generated by roxygen2 (4.0.1): do not edit by hand
\name{identify_covariates}
\alias{identify_covariates}
\title{identify_covariates}
\usage{
identify_covariates(covars, keep_n_covars = 200, indexes = FALSE)
}
\arguments{
\item{covars}{a matrix or something that can be coerced with \code{\link[base]{as.matrix}} of covariates}

\item{keep_n_covars}{number of covariates to keep}

\item{indexes}{Should indexes be returned? Or a subset of \code{covars}. Defaults to \code{FALSE}.}
}
\value{
Indexes of identified columns or a subset of \code{covars}
}
\description{
Given a matrix of covarites, \code{identify_covariates} returns the top \code{keep_n_covars} or the indexes of those columns.
}
\details{
The variance of each column of \code{covars} is calculated and sorted in descending order of column variances.

If \code{indexes==TRUE}, a vector of the top \code{keep_n_covars} column indexes is returned.

If \code{indexes==FALSE}, a matrix of covariates is returned whos columns are the top \code{keep_n_covars} colums of
\code{covars}. Columns are in their original order.
If also \code{keep_n_covars >= ncol(covar)}, then the function returns immediately without calculating column variances,
as it is unnecessary.

\strong{Differences from Schneeweiss et al. (2009):}
\itemize{
  \item{Instead of ranking covariates by prevalence (or equivalently, \code{var(col1 > 0)} as covariates are typically counts),
  covariates are ranked by variance.  This means that a covariates \code{col1=c(0,1,3)} will be ranked higher than \code{col2=c(0,1,1)}.
  If ranked by prevalence, \code{col1} and \code{col2} would be the same.
  }
  \item{Covariates that have fewer than 100 non-zero values are not automatically dropped.
  If typical covariates tend to have more than 100 non-zero values will typically be ranked higher than those with fewer than 100
  automatically.}
}
}
\author{
Sam Lendle
}
\references{
Schneeweiss, S., Rassen, J. A., Glynn, R. J., Avorn, J., Mogun,
H., & Brookhart, M. A. (2009). High-dimensional propensity score adjustment
in studies of treatment effects using health care claims data. \emph{Epidemiology
(Cambridge, Mass.)}, 20(4), 512.
}

