% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plot_bias_distribution}
\alias{plot_bias_distribution}
\title{Plot bias distribution for top covariates}
\usage{
plot_bias_distribution(hdps_result, top_n = 50, interactive = FALSE)
}
\arguments{
\item{hdps_result}{Results from prioritize function}

\item{top_n}{Number of top covariates to display}

\item{interactive}{Whether to create interactive plots}
}
\value{
ggplot or plotly object
}
\description{
Plot bias distribution for top covariates
}
