% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plot_bias_vs_prevalence}
\alias{plot_bias_vs_prevalence}
\title{Plot bias vs prevalence}
\usage{
plot_bias_vs_prevalence(hdps_result, interactive = FALSE)
}
\arguments{
\item{hdps_result}{Results from prioritize function}

\item{interactive}{Whether to create interactive plots}
}
\value{
ggplot or plotly object
}
\description{
Plot bias vs prevalence
}
