% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.hdps_covars}
\alias{predict.hdps_covars}
\title{Get matrix of hdps selected covariates}
\usage{
\method{predict}{hdps_covars}(object, newdata = NULL, keep_k_total)
}
\arguments{
\item{object}{object of class \code{hdps_covars}}

\item{newdata}{\code{NULL}, or a matrix who's columns have names corresponding to those selected by hdps in \code{object}.
If \code{NULL} selected covariates from original matrix used in the screening step are returned.}

\item{keep_k_total}{change \code{keep_k_total} from the original call to \code{\link{hdps_screen}}}
}
\value{
A matrix of hdps selected covariates
}
\description{
returns the matix of covariates based on an hdps screening
}
\details{
.. content for \details{} ..
}
\author{
Sam Lendle
}
\seealso{
\link{hdps_screen}
}

