% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hdps.R
\name{predict.SL_hdps}
\alias{predict.SL_hdps}
\title{Get predictions from SL_hdps wrapper}
\usage{
\method{predict}{SL_hdps}(object, newdata, ...)
}
\arguments{
\item{object}{object of class \code{SL_hdps}}

\item{newdata}{a matrix of covariates to predict from}

\item{...}{ignored}
}
\value{
vector of predictions
}
\description{
Get predictions from SL_hdps wrapper
}
\author{
Sam Lendle
}
