% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prioritize_covariates.R
\name{prioritize_covariates}
\alias{prioritize_covariates}
\title{prioritize_covariates}
\usage{
prioritize_covariates(outcome, treatment, covars, return_bias = FALSE,
  keep_NaNs = FALSE)
}
\arguments{
\item{outcome}{binary vector of outcomes}

\item{treatment}{binary vector of treatments}

\item{covars}{\code{matrix} or \code{data.frame} of binary covariates.}

\item{return_bias}{Should the calculated multiplitive biases be returned with the ordered indexes? Defaults to \code{FALSE}}

\item{keep_NaNs}{If the calculated multiplicitive bias for a covariate is \code{NaN}, should its index be included in the returned vector? Defaults to \code{FALSE}}
}
\value{
Vector of column indexes of \code{covars}.
}
\description{
Returns the indexes of columns \code{covars} ordered by the possible amount of counfounding each column could adjust for.
}
\details{
For each covariate in \code{covars}, the potential multiplicitive bias is calculated as described in Schneeweiss et al. (2009).
The column indexes are then put in descending order of the absolute value of the log of the multiplicitive bias.

If \code{return_bias==TRUE}, the returned vector of indexes includes the multiplicitive biases sorted in the same order and stored in an attribute called \code{"bias_m"}.

If \code{outcome} has no variation for a particular value of a covariate, then the multiplicitive bias is calculated as \code{NaN}.
If \code{keep_NaNs==FALSE}, then the column indexes of such covariates are not included in the returned vector.
}
\author{
Sam Lendle
}
\references{
Schneeweiss, S., Rassen, J. A., Glynn, R. J., Avorn, J., Mogun,
H., & Brookhart, M. A. (2009). High-dimensional propensity score adjustment
in studies of treatment effects using health care claims data. \emph{Epidemiology
(Cambridge, Mass.)}, 20(4), 512.
}

