% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{rec_assess}
\alias{rec_assess}
\title{covariate assessment}
\usage{
rec_assess(dt, id, code, type, rank = Inf)
}
\arguments{
\item{dt}{A database including the pid and covariates for assessment}

\item{id}{The column name of patient identification}

\item{code}{The column name of codes, eg. ICD-9, ICD-19, Drug Codes, or Lab results}

\item{type}{The prefix for output table, indicating type of data dimensions, eg. dx for diagnoses, px for procedures}

\item{rank}{The desired number of covariates to include in the model. eg. 500, 1000, Inf, etc}
}
\value{
A data.table including the recurrent assessment of the codes
}
\description{
The recurrence of features is assessed in a pre-exposure period, creating binary covariates based on a set of frequency-based cut-offs.The standard implementation of the HDPS defines three indicators for each patient capturing whether a feature was recorded:larger than once,larger than the median, and larger the75th percentile.
}
