% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SL.hdps.R
\name{screen}
\alias{screen}
\alias{screen.excludenames}
\alias{screen.names}
\title{SuperLearner screening wrappers}
\usage{
screen.names(names)

screen.excludenames(names)
}
\arguments{
\item{names}{Names to be included or excluded

These functions generate simple screening wrappers for SuperLearner to
include or exclude variables based on \code{names}. This is is helpful
because in order to use HDPS as a candidate in SuperLearner, you need to
include the study outcome variable as a covariate. But to use a non-HDPS
algorithm, (say a random forest on some specified set of covariates,) as a
candidate as well, you want to make sure you're not adjusting for the
outcome which is downstream from treatment on the causal pathway.

See documentation for the SuperLearner package for more about screening algorithms.}
}
\description{
SuperLearner screening wrappers
}
\examples{
screen.predefined <- screen.names(c("names", "of", "predefined",
"covariates", "that", "definitely", "dont", "include", "the", "outcome"))

screen.notoutcome <- screen.excludenames(c("outcome_variable_name",
"and", "other", "covariates", "to", "exclude"))
}

