% Generated by roxygen2 (4.0.2): do not edit by hand
\name{assess_recurrence}
\alias{assess_recurrence}
\title{assess_recurrence}
\usage{
assess_recurrence(covars)
}
\arguments{
\item{covars}{a matrix or something that can be coerced with \code{\link[base]{as.matrix}} of covariates}
}
\value{
Expanded \code{covars} matrix.
}
\description{
Expands covarites to up to three binary columns for at least one, sporadic, or frequent occrence of each covariate.
}
\details{
Each column \code{x} of \code{covars} is expanded to three binary columns.
The first column indicates that the value of \code{x} is non-zero.
The second indicates that the value of \code{x} \eqn{\ge} the median of non-zero values of \code{x}.
The third indicates that the value of \code{x} \eqn{\ge} the 75th percentile of non-zero values of \code{x}.
Non-unique columns per covariate are dropped.

Groups of columns of the returned matrix are in the same order of columns in \code{covars}.

If \code{covars} has column names, the returned matrix will have the same column names with suffexes \code{"_once"},
\code{"_sporadic"}, and \code{"_frequent"} for the first, second, and third expanded columns, respectively.
}
\author{
Sam Lendle
}
\references{
Schneeweiss, S., Rassen, J. A., Glynn, R. J., Avorn, J., Mogun,
H., & Brookhart, M. A. (2009). High-dimensional propensity score adjustment
in studies of treatment effects using health care claims data. \emph{Epidemiology
(Cambridge, Mass.)}, 20(4), 512.
}

