% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{estBias}
\alias{estBias}
\title{Covariate assessment wrapper function}
\usage{
estBias(hdpsCohort, cova, expo, outc, correction = TRUE)
}
\arguments{
\item{hdpsCohort}{Dataset after recurrence assessment}

\item{cova}{Column name for covariate}

\item{expo}{Column name for exposure}

\item{outc}{Column name for outcome}

\item{correction}{Apply 0.1 correction for zero cells}
}
\value{
Data table with bias estimates
}
\description{
Combines identify_candidates and assess_recurrence for backward compatibility.
Estimate apparent relative risk bias for a covariate
}
