% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{estBias}
\alias{estBias}
\title{ARR Bias estimation for all covariates}
\usage{
estBias(hdpsCohort, cova, expo, outc, correction = TRUE, ...)
}
\arguments{
\item{hdpsCohort}{The dataset after recurrent assesemnt in data.table format}

\item{cova}{Column name of the desire covariates}

\item{expo}{Column name of the desire exposure}

\item{outc}{Column name of the desire outcome}

\item{correction}{When the outcome is rare, one of the 2 by 2 table could be zero. In this case, the algorithm will add 0.1 to the cell.}

\item{...}{}
}
\value{
A data.table including the apparent relative risk for one covariates
}
\description{
ARR Bias estimation for all covariates
}
