% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{estBiasTable}
\alias{estBiasTable}
\title{Estimate the apparent relative risk for one covariate}
\usage{
estBiasTable(dt, expo, cova, ...)
}
\arguments{
\item{dt}{A dataset in data.table format, including}

\item{expo}{The column name for exposure or outcome}

\item{cova}{The column name for covariates}

\item{...}{}
}
\value{
A data.table including the apparent relative risk for one covariate
}
\description{
Estimate the apparent relative risk for one covariate
}
