% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdps_screen.R
\name{hdps_screen}
\alias{hdps_screen}
\title{hdps_screen}
\usage{
hdps_screen(outcome, treatment, covars, dimension_names = NULL,
  dimension_indexes = NULL, keep_n_per_dimension = 200,
  keep_k_total = 500, verbose = FALSE)
}
\arguments{
\item{outcome}{binary vector of outcomes}

\item{treatment}{binary vector of treatments}

\item{covars}{\code{matrix} or \code{data.frame} of binary covariates.}

\item{dimension_names}{A character vector of patterns to match against the column names of \code{covars} to split columns into dimension groups. See details.}

\item{dimension_indexes}{A list of vectors of column indexes corresponding to dimension groups. See details. Cannot be specified with \code{dimension_names}.}

\item{keep_n_per_dimension}{The maximum number of covariates to be kept per dimension by \code{\link{identify_covariates}}.}

\item{keep_k_total}{Total number of covariates to keep after expanding by \code{\link{assess_recurrence}} and ordering by \code{link{prioritize_covariates}}.}

\item{verbose}{Should verbose output be printed?}
}
\value{
An object of class \code{hdps_covars}
}
\description{
The \code{hdps_screen} function performs part of step 2 (\code{\link{identify_covariates}}),
steps 3 (\code{\link{assess_recurrence}}) and 4 (\code{\link{prioritize_covariates}})
of the HDPS algorithm (Schneeweiss et al., 2009).
}
\details{
The \code{hdps_screen} function performs part of step 2 (\code{\link{identify_covariates}}),
steps 3 (\code{\link{assess_recurrence}}) and 4 (\code{\link{prioritize_covariates}})
of the HDPS algorithm (Schneeweiss et al., 2009).

\emph{Step 2.} Columns of \code{covars} are split by data dimension (as defined in Schneeweiss et al. (2009)) and
filtered by \code{\link{identify_covariates}}.

Dimensions can be specified in two ways.
If \code{dimension_names} is used, the \code{colnames(covars)} is \code{\link[base]{grep}}ed for each value of
\code{dimension_names}.
If some column names match more than one pattern, an error is thrown.
If some column names are not matched by any pattern, a warning is issued and those columns are ignored.
For example, suppose the column names of \code{covars} are \code{c("drug_1", "drug_2", "proc_1", "proc_2")}.
\code{dimension_names <- c("drug", "proc")} would split \code{covars} into two dimensions,
one for \code{drug}s and one for \code{proc}s.

Dimensions can also be specified by \code{dimension_indexes} which should contain a list of either column
indexes or column names for each dimension.

If neither \code{dimension_names} nor \code{dimension_indexes} is specified, all covariates are treated as one dimension.

\emph{Step 3.} After filtering, remaining covariates are expanded by \code{\link{assess_recurrence}}.

If at this point, the number of expanded covariates is less than \code{keep_k_total}, all expanded covariates are returned.

\emph{Step 4.} Expanded covariates are ordered with \code{\link{prioritize_covariates}}.

\emph{Step 5.} Step 5 can be performed with \code{\link{predict.hdps_covars}}.
}
\examples{
set.seed(123)
n <- 1000
p <- 10000
out <- rbinom(n, 1, 0.05)
trt <- rbinom(n, 1, 0.5)
covars <- matrix(rbinom(n*p, 3, 0.05), n)
colnames(covars) <- c(paste("drug", 1:(p/2), sep="_"),
                      paste("proc", 1:(p/2), sep="_"))

dimension_names <- c("drug", "proc")

screened_covars_fit <- hdps_screen(out, trt, covars,
                                   dimension_names = dimension_names,
                                   keep_n_per_dimension = 400,
                                   keep_k_total = 200,
                                   verbose=TRUE)

screened_covars <- predict(screened_covars_fit)
}
\author{
Sam Lendle
}
\references{
Schneeweiss, S., Rassen, J. A., Glynn, R. J., Avorn, J., Mogun,
H., & Brookhart, M. A. (2009). High-dimensional propensity score adjustment
in studies of treatment effects using health care claims data. \emph{Epidemiology
(Cambridge, Mass.)}, 20(4), 512.
}
\seealso{
\code{\link{predict.hdps_covars}}
}

