% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plot_covariate_strength}
\alias{plot_covariate_strength}
\title{Plot covariate strength relationships}
\usage{
plot_covariate_strength(hdps_result, interactive = FALSE)
}
\arguments{
\item{hdps_result}{Results from prioritize function}

\item{interactive}{Whether to create interactive plots}
}
\value{
ggplot or plotly object
}
\description{
Plot covariate strength relationships
}
