% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SL.hdps.R
\name{SL.hdps.generator}
\alias{SL.hdps.generator}
\title{SL.hdps.generator}
\usage{
SL.hdps.generator(out_name, dimension_names, predef_covar_names = c(),
  keep_k_total, ...)
}
\arguments{
\item{out_name}{Name of the outcome variable.}

\item{dimension_names}{Dimension names of HDPS dimensions. See
\code{\link{hdps_screen}}.}

\item{predef_covar_names}{Names of predefined covariates to be included in
logistic regression model.}

\item{keep_k_total}{See \code{\link{hdps_screen}}.}

\item{...}{Other arguments passed to \code{\link{hdps_screen}}.}
}
\value{
A SuperLearner wrapper function
}
\description{
Generates a wrapper for SuperLearner using HDPS
}
\details{
A HDPS candidate will generate covariates using \code{hdps_screen} from
codes, and estimate the propensity score with logistic regression on
generated covariates and predefined covariates.

To use HDPS in SuperLearner to estimate a propensity score, you need to
include the outcome variable as a covariate where here outcome means the
outcome of interest in the causal problem as opposed to the \code{Y}
variable in SuperLearner. For non-HDPS candidates in SuperLearner, it's
important to exclude the outcome variable via \code{\link[=screen]{screen.named}} or
some other screening algorithm in order to avoid adjusting for something
downstream on the causal pathway.
}
\author{
Sam Lendle
}

